+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 2.5.8
+* Fixed missing translation keys
+* Added hide buy stack in merchant settings
+* Added option to change max value for buy from merchant in merchant settings
+
 ## Version 2.5.7
 * WFRP fixed generator text
 * WFRP Changed currency to localized currency